requirements.txt
# FastAPI and ASGI server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.36
alembic==1.13.1
pymysql==1.1.0
cryptography>=41.0.0  # Required for MySQL SSL connections

# Security and Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

# Configuration
pydantic>=2.5.0
pydantic-settings>=2.1.0
email-validator>=2.0.0

# Template engine
jinja2==3.1.3

# File processing
PyPDF2==3.0.1
python-docx==0.8.11

# NLP and Plagiarism Detection
nltk>=3.8.1
scikit-learn>=1.3.0
numpy>=1.24.0

# External source checking
googlesearch-python>=1.2.0

# Utilities
python-dotenv==1.0.0
